# pet-sitter-server/openapi.yml
openapi: 3.0.3
info:
  title: PetSitter API
  version: "0.1"
...
paths:
...
  /photos/upload:
    post:
      tags:
      - Photos
      summary: Upload Multiple Photos (Pre-upload)
      description: |
        여러 파일을 한 번에 업로드합니다. entity와 미연결 상태(orphan)로 저장됩니다.
        반환된 Photo IDs를 POST /jobs 등의 요청 body에 포함해 연결하세요.
      operationId: uploadPhotos
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - files
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 10
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photo'
        '400':
          description: Bad Request (invalid file type or size)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
  /users/{id}/photos:
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    post:
      tags:
      - Photos
      summary: Upload User Profile Photo
      operationId: uploadUserPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Bad Request (invalid file type or size)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
  /jobs/{id}/photos:
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    post:
      tags:
      - Photos
      summary: Upload Job Photo
      operationId: uploadJobPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Bad Request (invalid file type or size)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
  /pets/{id}/photos:
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    post:
      tags:
      - Photos
      summary: Upload Pet Photo
      operationId: uploadPetPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photo'
        '400':
          description: Bad Request (invalid file type or size)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
        '404':
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
  /photos/{id}:
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
    delete:
      tags:
      - Photos
      summary: Delete Photo
      operationId: deletePhoto
      responses:
        '204':
          description: No Content
        '403':
          description: Forbidden (not the uploader)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
        '404':
          description: Photo not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OASError'
...
components:
  schemas:
...
    Photo:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        url:
          type: string
          description: Publicly accessible URL for the photo
        file_name:
          type: string
          description: Stored file name (UUID-based)
        original_name:
          type: string
          description: Original uploaded file name
        mime_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
        size:
          type: integer
          description: File size in bytes
        uploader_id:
          type: string
          readOnly: true
        user_id:
          type: string
          nullable: true
          description: Profile photo target user ID
        job_id:
          type: string
          nullable: true
          description: Associated job ID
        pet_id:
          type: string
          nullable: true
          description: Associated pet ID
        createdAt:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - url
        - file_name
        - original_name
        - mime_type
        - size
        - uploader_id
        - createdAt
    OASError:
      type: object
      properties:
        message:
          type: string
          description: Human-readable error message
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                description: For input validation errors, identifies where in the JSON request body the error occured
              message:
                type: string
                description: Human-readable error message
              errorCode:
                type: string
                description: Code indicating error type
...
  securitySchemes:
    SessionToken:
      type: apiKey
      in: header
      name: Authorization
security:
  - SessionToken: []